Microsoft paper
* [Large-Scale Intelligent Microservices](https://arxiv.org/pdf/2009.08044.pdf) - Microsoft paper that an Apache Spark-based micro-service orchestration framework that extends database operations to include web service primitives.

Abstract—Deploying Machine Learning (ML) algorithms
within databases is a challenge due to the varied computational
footprints of modern ML algorithms and the myriad of database
technologies each with their own restrictive syntax. We introduce
an Apache Spark-based micro-service orchestration framework
that extends database operations to include web service primitives. 
Our system can orchestrate web services across hundreds
of machines and takes full advantage of cluster, thread, and
asynchronous parallelism. Using this framework, we provide
large scale clients for intelligent services such as speech, vision,
search, anomaly detection, and text analysis. This allows users
to integrate ready-to-use intelligence into any datastore with an
Apache Spark connector. To eliminate the majority of overhead
from network communication, we also introduce a low-latency
containerized version of our architecture. Finally, we demonstrate
that the services we investigate are competitive on a variety of
benchmarks, and present two applications of this framework to
create intell